changed MATRIX_WIDTH HEIGHT to negatives
LEDMatrix uses negative values to reverse direction for MATRIX_WIDTH & HEIGHT. My generator used sepeate defines. I changed these back to the LEDMatrix convension of negatives.
My added feature to be able to reverse the led direction inside tiles is still available for the table only. It is not in the library code